[DennisLee][L6][Calculator] Split the test.py file into unit-test.py and integration-test.py. Ready to submit.
