+++ students/DennisLee/lesson06/unit-test.py
@@ -158,51 +158,3 @@ class CalculatorTests(TestCase):
         self.calculator.enter_number(0)
         result = self.calculator.divide()
         self.assertEqual(result, 0)
-
-
-class ModuleTests(TestCase):
-    """Test the calculator module/UI during random, sustained usage."""
-    def setUp(self):
-        self.calculator = Calculator(
-            Adder(), Subtracter(), Multiplier(), Divider())
-
-    def test_module(self):
-        """
-        Make sure the calculator returns the correct value after a
-        series of random number entries and operations.
-        """
-        self.calculator.enter_number(5)
-        self.calculator.enter_number(2)
-
-        self.calculator.multiply()
-
-        self.calculator.enter_number(46)
-
-        self.calculator.add()
-
-        self.calculator.enter_number(8)
-
-        self.calculator.divide()
-
-        self.calculator.enter_number(1)
-
-        result = self.calculator.subtract()
-
-        self.assertEqual(6, result)
-
-    def test_after_division_by_zero(self):
-        """
-        Make sure an insufficient operands exception occurs if a user
-        immediately invokes an operation after a division by zero error
-        without specifying another number. (Following a division by
-        zero, the stack is cleared, and only a zero is left there.)
-        """
-        self.calculator.enter_number(15)
-        self.calculator.enter_number(5)
-        self.calculator.enter_number(0)
-        self.calculator.divide()
-        with self.assertRaises(InsufficientOperands):
-            self.calculator.subtract()
-        self.calculator.enter_number(8)
-        result = self.calculator.subtract()
-        self.assertEqual(-8, result)
